Conflicts:
	ProjectFlightSchool/Graphics/Graphics.cpp
	ProjectFlightSchool/Graphics/Graphics.h
	ProjectFlightSchool/Graphics/Graphics.vcxproj
	ProjectFlightSchool/Graphics/Graphics.vcxproj.filters
	ProjectFlightSchool/ProjectFlightSchool/PlayState.h
	ProjectFlightSchool/ProjectFlightSchool/RenderManager.cpp
	ProjectFlightSchool/ProjectFlightSchool/RenderManager.h
	ProjectFlightSchool/ProjectFlightSchool/RenderStructs.h

+++ ProjectFlightSchool/ProjectFlightSchool/ParticleManager.h
@@ -15,11 +15,11 @@ class ParticleManager
 {
 	private:
 		ParticleSystem***	mParticleSystems;
-		size_t				mNrOfParticleSystems;
-		size_t				mNrOfActiveParticleSystems;
-		size_t*				mNrOfParticleSystemsPerType;
-		size_t*				mMaxNrOfParticleSystemsPerType;
-		size_t*				mNrOfActiveParticleSystemsPerType;
+		int					mNrOfParticleSystems;
+		int					mNrOfActiveParticleSystems;
+		int*				mNrOfParticleSystemsPerType;
+		int*				mMaxNrOfParticleSystemsPerType;
+		int*				mNrOfActiveParticleSystemsPerType;
 		
 
 	public:
